Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
- int radius;
- scanf("%d",&raduis);
- printf("Area: %d\n",3.14*radius*radius);
+ float radius;
+ scanf("%f",&raduis);
+ printf("Area: %f\n",3.14*radius*radius);
  return 0;
 }
